pyproject.toml
[project]
name = "get-wrapped"
version = "0.1.3"
description = "Create your own Wrapped-style summary from any dataset"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.32.2",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
authors = [{name = "Kushagra Kumar"}]
license = {text = "MIT"}
homepage = "https://github.com/kkumar/get-wrapped"


[build-system]
requires = ["uv_build >= 0.9.17, <0.10.0"]
build-backend = "uv_build"
